+++ core/badge-urls/make-badge-url.js
@@ -59,6 +59,7 @@ function staticBadgeUrl({
   baseUrl = '',
   label,
   message,
+  labelColor,
   color = 'lightgray',
   style,
   namedLogo,
@@ -66,6 +67,7 @@ function staticBadgeUrl({
 }) {
   const path = [label, message, color].map(encodeField).join('-')
   const outQueryString = queryString.stringify({
+    labelColor,
     style,
     logo: namedLogo,
   })